+++ tests/unit/connection.ts
@@ -687,6 +687,20 @@ describe('Connection', function() {
           conn.disconnect();
           sinon.assert.calledOnce(mediaStream.close);
         });
+
+        [
+          'answer',
+          'cancel',
+          'hangup',
+          'ringing',
+          'transportClose',
+        ].forEach((eventName: string) => {
+          it(`should call pstream.removeListener on ${eventName}`, () => {
+            conn.disconnect();
+            clock.tick(10);
+            assert.equal(pstream.listenerCount(eventName), 0);
+          });
+        });
       });
     });